Move the files from the certificate folder to the invoice folder
