+++ src/components/invoice/input_invoice.tsx
@@ -1,20 +1,20 @@
 import React from 'react';
 import Pagination from '@/components/pagination/pagination';
 import { DISPLAY_LIST_VIEW_TYPE } from '@/constants/display';
-import OutputInvoiceTable from '@/components/certificate/output_invoice_table';
+import InputInvoiceTable from '@/components/invoice/input_invoice_table';
 import CertificateGrid from '@/components/certificate/certificate_grid';
 import { SortOrder } from '@/constants/sort';
 import { CertificateTab } from '@/constants/certificate';
 import { CurrencyType } from '@/constants/currency';
-import { ICertificateRC2OutputUI } from '@/interfaces/certificate_rc2';
+import { ICertificateRC2InputUI } from '@/interfaces/certificate_rc2';
 
-interface OutputInvoiceProps {
+interface InputInvoiceProps {
   activeTab: CertificateTab;
   page: number;
   setPage: React.Dispatch<React.SetStateAction<number>>;
   totalCount: number;
   totalPages: number;
-  certificates: ICertificateRC2OutputUI[]; // Info: (20240923 - Anna) 項目列表
+  certificates: ICertificateRC2InputUI[]; // Info: (20240923 - Anna) 項目列表
   currencyAlias: CurrencyType;
   viewType?: DISPLAY_LIST_VIEW_TYPE; // Info: (20240923 - Anna) 顯示模式
   activeSelection: boolean; // Info: (20240923 - Anna) 是否處於選擇狀態
@@ -27,17 +27,18 @@ interface OutputInvoiceProps {
   dateSort: SortOrder | null;
   amountSort: SortOrder | null;
   voucherSort: SortOrder | null;
-  certificateNoSort: SortOrder | null; // Info: (20250416 - Anna) 憑證號碼排序
   certificateTypeSort: SortOrder | null; // Info: (20250416 - Anna) 憑證類型排序
+  certificateNoSort: SortOrder | null;
   setDateSort: React.Dispatch<React.SetStateAction<SortOrder | null>>;
   setAmountSort: React.Dispatch<React.SetStateAction<SortOrder | null>>;
   setVoucherSort: React.Dispatch<React.SetStateAction<SortOrder | null>>;
-  setCertificateNoSort: React.Dispatch<React.SetStateAction<SortOrder | null>>; // Info: (20250416 - Anna) 憑證號碼排序
   setCertificateTypeSort: React.Dispatch<React.SetStateAction<SortOrder | null>>; // Info: (20250416 - Anna) 憑證類型排序
+  setCertificateNoSort: React.Dispatch<React.SetStateAction<SortOrder | null>>;
+  isExporting: boolean;
 }
 
 // Deprecated: (20240919 - Anna) will be replaced by actual data type
-const OutputInvoice: React.FC<OutputInvoiceProps> = ({
+const InputInvoice: React.FC<InputInvoiceProps> = ({
   activeTab,
   page,
   setPage,
@@ -61,13 +62,14 @@ const OutputInvoice: React.FC<OutputInvoiceProps> = ({
   setDateSort,
   setAmountSort,
   setVoucherSort,
-  setCertificateNoSort,
   setCertificateTypeSort,
+  setCertificateNoSort,
+  isExporting,
 }) => {
   return (
     <>
       {viewType === DISPLAY_LIST_VIEW_TYPE.LIST && (
-        <OutputInvoiceTable
+        <InputInvoiceTable
           activeTab={activeTab}
           certificates={certificates}
           currencyAlias={currencyAlias}
@@ -84,8 +86,9 @@ const OutputInvoice: React.FC<OutputInvoiceProps> = ({
           setDateSort={setDateSort}
           setAmountSort={setAmountSort}
           setVoucherSort={setVoucherSort}
-          setCertificateNoSort={setCertificateNoSort}
           setCertificateTypeSort={setCertificateTypeSort}
+          setCertificateNoSort={setCertificateNoSort}
+          isExporting={isExporting}
         />
       )}
       {viewType === DISPLAY_LIST_VIEW_TYPE.GRID && (
@@ -113,4 +116,4 @@ const OutputInvoice: React.FC<OutputInvoiceProps> = ({
   );
 };
 
-export default OutputInvoice;
+export default InputInvoice;
