+++ src/components/invoice/input_invoice_table.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { useTranslation } from 'next-i18next';
-import OutputInvoiceItem from '@/components/certificate/output_invoice_item';
+import InputInvoiceItem from '@/components/invoice/input_invoice_item';
 import SortingButton from '@/components/voucher/sorting_button';
 import { SortOrder } from '@/constants/sort';
 import { HiCheck } from 'react-icons/hi';
 import { CertificateTab } from '@/constants/certificate';
 import { CurrencyType } from '@/constants/currency';
-import { ICertificateRC2OutputUI } from '@/interfaces/certificate_rc2';
+import { ICertificateRC2InputUI } from '@/interfaces/certificate_rc2';
 
-interface OutputInvoiceTableProps {
+interface InputInvoiceTableProps {
   activeTab: CertificateTab;
-  certificates: ICertificateRC2OutputUI[];
+  certificates: ICertificateRC2InputUI[];
   currencyAlias: CurrencyType;
   activeSelection: boolean; // Info: (20240923 - Anna) 是否處於選擇狀態 // Info: (20240923 - Anna) 選中的項目 ID 列表
   handleSelect: (ids: number[], isSelected: boolean) => void; // Info: (20240923 - Anna) 當選擇變更時的回調函數
@@ -20,16 +20,17 @@ interface OutputInvoiceTableProps {
   dateSort: SortOrder | null;
   amountSort: SortOrder | null;
   voucherSort: SortOrder | null;
-  certificateNoSort: SortOrder | null; // Info: (20250416 - Anna) 憑證號碼排序
+  certificateNoSort: SortOrder | null;
   certificateTypeSort: SortOrder | null; // Info: (20250416 - Anna) 憑證類型排序
   setDateSort: React.Dispatch<React.SetStateAction<SortOrder | null>>;
   setAmountSort: React.Dispatch<React.SetStateAction<SortOrder | null>>;
   setVoucherSort: React.Dispatch<React.SetStateAction<SortOrder | null>>;
-  setCertificateNoSort: React.Dispatch<React.SetStateAction<SortOrder | null>>; // Info: (20250416 - Anna) 憑證號碼排序
   setCertificateTypeSort: React.Dispatch<React.SetStateAction<SortOrder | null>>; // Info: (20250416 - Anna) 憑證類型排序
+  setCertificateNoSort: React.Dispatch<React.SetStateAction<SortOrder | null>>;
+  isExporting: boolean;
 }
 
-const OutputInvoiceTable: React.FC<OutputInvoiceTableProps> = ({
+const InputInvoiceTable: React.FC<InputInvoiceTableProps> = ({
   activeTab,
   activeSelection,
   certificates,
@@ -46,8 +47,9 @@ const OutputInvoiceTable: React.FC<OutputInvoiceTableProps> = ({
   setDateSort,
   setAmountSort,
   setVoucherSort,
-  setCertificateNoSort,
   setCertificateTypeSort,
+  setCertificateNoSort,
+  isExporting,
 }) => {
   const { t } = useTranslation('certificate');
   const displayedIssuedDate = SortingButton({
@@ -57,8 +59,8 @@ const OutputInvoiceTable: React.FC<OutputInvoiceTableProps> = ({
     handleReset: () => {
       setAmountSort(null);
       setVoucherSort(null);
-      setCertificateNoSort(null);
       setCertificateTypeSort(null);
+      setCertificateNoSort(null);
     },
   });
 
@@ -69,8 +71,8 @@ const OutputInvoiceTable: React.FC<OutputInvoiceTableProps> = ({
     handleReset: () => {
       setDateSort(null);
       setVoucherSort(null);
-      setCertificateNoSort(null);
       setCertificateTypeSort(null);
+      setCertificateNoSort(null);
     },
   });
 
@@ -81,12 +83,24 @@ const OutputInvoiceTable: React.FC<OutputInvoiceTableProps> = ({
     handleReset: () => {
       setDateSort(null);
       setAmountSort(null);
-      setCertificateNoSort(null);
       setCertificateTypeSort(null);
+      setCertificateNoSort(null);
     },
   });
 
-  // Info: (20250416 - Anna) 憑證號碼表頭
+  // Info: (20250416 - Anna) 憑證類型表頭
+  const displayedInvoiceType = SortingButton({
+    string: t('certificate:TABLE.INVOICE_TYPE'),
+    sortOrder: certificateTypeSort,
+    setSortOrder: setCertificateTypeSort,
+    handleReset: () => {
+      setDateSort(null);
+      setAmountSort(null);
+      setVoucherSort(null);
+      setCertificateNoSort(null);
+    },
+  });
+
   const displayedCertificateNo = SortingButton({
     string: t('certificate:TABLE.INVOICE_NUMBER'),
     sortOrder: certificateNoSort,
@@ -96,19 +110,6 @@ const OutputInvoiceTable: React.FC<OutputInvoiceTableProps> = ({
       setAmountSort(null);
       setVoucherSort(null);
       setCertificateTypeSort(null);
-    },
-  });
-
-  // Info: (20250416 - Anna) 憑證類型表頭
-  const displayedCertificateType = SortingButton({
-    string: t('certificate:TABLE.INVOICE_TYPE'),
-    sortOrder: certificateTypeSort,
-    setSortOrder: setCertificateTypeSort,
-    handleReset: () => {
-      setDateSort(null);
-      setAmountSort(null);
-      setVoucherSort(null);
-      setCertificateNoSort(null);
     },
   });
 
@@ -128,30 +129,29 @@ const OutputInvoiceTable: React.FC<OutputInvoiceTableProps> = ({
                 </div>
               </div>
             )}
-            <div className="table-cell w-100px min-w-100px border-b border-r border-stroke-neutral-quaternary px-lv-2 text-center align-middle">
-              {displayedIssuedDate}
+            <div className="download-pb-4 table-cell w-100px min-w-100px border-b border-r border-stroke-neutral-quaternary px-lv-2 text-center align-middle">
+              {isExporting ? t('certificate:TABLE.DATE') : displayedIssuedDate}
             </div>
-            <div className="table-cell w-120px min-w-120px flex-col items-center border-b border-r border-stroke-neutral-quaternary px-lv-2 text-left align-middle">
-              {displayedCertificateNo}
+            <div className="download-pb-4 table-cell w-120px min-w-120px flex-col items-center border-b border-r border-stroke-neutral-quaternary px-lv-2 text-center align-middle">
+              {isExporting ? t('certificate:TABLE.INVOICE_NUMBER') : displayedCertificateNo}
             </div>
-            <div className="col-span-full table-cell min-w-100px border-b border-r border-stroke-neutral-quaternary px-lv-2 text-center align-middle">
-              {displayedCertificateType}
+            <div className="download-pb-4 col-span-full table-cell min-w-100px border-b border-r border-stroke-neutral-quaternary px-lv-2 text-center align-middle">
+              {isExporting ? t('certificate:TABLE.INVOICE_TYPE') : displayedInvoiceType}
             </div>
-
-            <div className="table-cell w-100px min-w-100px border-b border-r border-stroke-neutral-quaternary px-lv-2 text-center align-middle">
+            <div className="download-pb-4 table-cell w-100px min-w-100px border-b border-r border-stroke-neutral-quaternary px-lv-2 text-center align-middle">
+              {t('certificate:TABLE.DEDUCTION_TYPE')}
+            </div>
+            <div className="download-pb-4 table-cell w-100px min-w-100px border-b border-r border-stroke-neutral-quaternary px-lv-2 text-center align-middle">
               {t('certificate:TABLE.TAX')}
             </div>
-            <div className="col-span-full table-cell min-w-100px flex-col items-center border-b border-r border-stroke-neutral-quaternary px-lv-2 text-left align-middle">
-              <div>{t('certificate:TABLE.COUNTERPARTY')}</div>
-              <div>
-                ({t('certificate:COUNTERPARTY.CLIENT')}/{t('certificate:COUNTERPARTY.SUPPLIER')})
-              </div>
+            <div className="download-pb-4 col-span-full table-cell min-w-100px flex-col items-center border-b border-r border-stroke-neutral-quaternary px-lv-2 text-center align-middle">
+              {t('certificate:EDIT.SELLER')}
             </div>
-            <div className="table-cell w-170px min-w-170px border-b border-r border-stroke-neutral-quaternary px-lv-2 text-center align-middle">
-              {displayedAmount}
+            <div className="download-pb-4 table-cell w-170px min-w-170px border-b border-r border-stroke-neutral-quaternary px-lv-2 text-center align-middle">
+              {isExporting ? t('certificate:TABLE.AMOUNT') : displayedAmount}
             </div>
             <div className="table-cell w-120px min-w-120px flex-col items-center border-b border-stroke-neutral-quaternary px-lv-2 text-center align-middle">
-              <div>{t('certificate:TABLE.UPLOADER')}</div>
+              <div className="download-pb-4">{t('certificate:TABLE.UPLOADER')}</div>
               {activeTab === CertificateTab.WITH_VOUCHER && displayedVoucherNumber}
             </div>
           </div>
@@ -160,7 +160,7 @@ const OutputInvoiceTable: React.FC<OutputInvoiceTableProps> = ({
         <div className="table-row-group">
           {/* Deprecated: (20240919 - Anna) Example of dynamic rows, should map actual data here */}
           {certificates.map((certificate, index) => (
-            <OutputInvoiceItem
+            <InputInvoiceItem
               currencyAlias={currencyAlias}
               activeSelection={activeSelection}
               handleSelect={handleSelect}
@@ -175,4 +175,4 @@ const OutputInvoiceTable: React.FC<OutputInvoiceTableProps> = ({
   );
 };
 
-export default OutputInvoiceTable;
+export default InputInvoiceTable;
