+++ src/pages/users/accounting/input_invoice_list.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'next-i18next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { ILocale } from '@/interfaces/locale';
 import Head from 'next/head';
-import InputInvoiceListBody from '@/components/certificate/input_invoice_list_body';
+import InputInvoiceListBody from '@/components/invoice/input_invoice_list_body';
 import Layout from '@/components/beta/layout/layout';
 
 const CertificateListPage: React.FC = () => {
